requirements.txt
pygame==2.6.1
numpy==1.24.4
pyyaml==6.0.1
pillow==10.2.0
